Module Operator.Entry

Entry Operations

Module for accessing directory entry information.

val path : Opendal_core.Operator.entry -> string

path entry gets the full path of the directory entry.

@example

printf "Full path: %s\n" (Entry.path entry)

val name : Opendal_core.Operator.entry -> string

name entry gets the name (filename) of the directory entry.

@example

printf "Name: %s\n" (Entry.name entry)

val metadata : Opendal_core.Operator.entry -> Opendal_core.Operator.metadata

metadata entry gets the metadata for the directory entry.

@example

  let meta = Entry.metadata entry in
  printf "Size: %Ld\n" (Metadata.content_length meta)