Module Operator.OperatorInfo

Operator Information

Module for accessing operator configuration and capabilities.

val name : Opendal_core.Operator.operator_info -> string

name info gets the name of the operator instance.

val scheme : Opendal_core.Operator.operator_info -> string

scheme info gets the storage scheme (e.g., "fs", "s3", "gcs").

@example

printf "Using scheme: %s\n" (OperatorInfo.scheme info)

val root : Opendal_core.Operator.operator_info -> string

root info gets the root path configured for the operator.

@example

printf "Root path: %s\n" (OperatorInfo.root info)

val capability : Opendal_core.Operator.operator_info -> Opendal_core.Operator.capability

capability info gets the capability information for the operator.

@example

  let cap = OperatorInfo.capability info in
  if Capability.write cap then print_endline "Write operations supported"